+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,11 +1,10 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.adapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.adapter.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mapping.HandlerMappingRegistry;
-import com.interface21.webmvc.servlet.view.JspView;
+import com.interface21.webmvc.servlet.view.ModelAndView;
+import com.interface21.webmvc.servlet.view.View;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
