+++ mvc/src/main/java/com/interface21/webmvc/servlet/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
 package com.interface21.webmvc.servlet.adapter;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
