+++ mvc/src/main/java/com/interface21/webmvc/servlet/stereotype/Controller.java
@@ -1,0 +1,14 @@
+package com.interface21.webmvc.servlet.stereotype;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Controller {
+    String value() default "";
+
+    String path() default "";
+}
